enable installation check for driver.sys
For now we still only support subfunction 3 (return drive table) and subfunction 0 (check if installed) which we say yes.  This allows DOS Internals DRVINFO.EXE to correctly use internal drive table (it checks if installed and doesn't bother if not) and return corresponding information.  Note: subfunction 3 is still handled in int2f.asm for now, but may be moved to cleaner (but bigger and slower) implementation in inthndlr.c