CI: Downcase TARGET variable
The main reason for this is so the filenames are downcased
on Linux systems such as the CI, which has the benefit of not
creating mixed case filenames within the Actions output zip.
On Windows and DOS systems there will be no appreciable
difference seen as fileystems there are either case
insensitive or case preserving.