disk: Set volume label also on BPB update
The existing code is only designed to update the serial number, but
MS-DOS and DR-DOS also update the volume and filesystem type too.

This patch enables updating of the volume field which is the
common case, but in the future it may be considered to update the
filesystem type also. DOS provides two methods of updating the
BPB (int21/6901 and int21/440d/0846), on FreeDOS the same code
implements both functions.

I have added tests on Dosemu2's test suite to cover serial number,
volume and filesystem type, the latter is marked as unsupported.